splitting info files in progress
