+++ myStyles.js
@@ -1,6 +1,6 @@
 import { StyleSheet, StatusBar } from 'react-native';
 
-export const styles = StyleSheet.create({
+const styles = StyleSheet.create({
   container: {
     flex: 1,
     marginTop: StatusBar.currentHeight || 0,
@@ -56,3 +56,5 @@ export const styles = StyleSheet.create({
     borderWidth: 1,
   },
 });
+
+export default styles;
